# docs/api.yml
swagger: "2.0" # Swagger version compatible with OAS 3.0
info:
 description: "De API van Bron"
 version: "1.0.0" # This is your api version
 title: "Bron API"
host: "api.bron.live"
basePath: "/"
paths:
  /users/forgot-password:
    post:
      tags:
        - Users
      summary: "Send a forgotten password email to a user"
      description: "This starts a forgotten password sequence for the specified user. A link will be sent where the user can reset his/hr password."
      produces:
        - "application/json"
      parameters:
      - name: "email"
        in: "form" # This is saying that this parameter must be given in a query string
        description: "The email address for which a forgotten password email is sent."
        type: "String"
        required: true
        example: "breyten@openstate.eu"
      responses:
        301:
          description: "The email was succesfully sent. Redirect to front page."
        400:
          description: "Something went wrong sending out the password lost email"
          schema:
            type: "object"
            error:
              type: "string"
              example: "Something wemt wrong."
          examples:
            application/json: {"error": "Something went wrong."}
  /users/change-password:
    post:
      tags:
        - Users
      summary: "Change a password following forgotten password"
      description: "This changes the password of the user. The user must have forgotten their password and must provide the new password."
      produces:
        - "application/json"
      parameters:
      - name: "changePasswordId"
        in: "form"
        description: "The token provided in the forget password e-mail."
        type: "String"
        required: true
        example: "6e11ceb9-2b26-4c40-b399-182d4f9b97b6"
      - name: "password"
        in: "form"
        description: "The new password."
        type: "String"
        required: true
        example: "password123"
      - name: "password_repeat"
        in: "form"
        description: "The new password repeated."
        type: "String"
        required: true
        example: "password123"
      responses:
        200:
          description: "The password was changed successfully."
          schema:
            type: "object"
            success:
              type: "string"
          examples:
            application/json: {"success":"true"}
        400:
          description: "Something went wrong changing the password."
          schema:
            type: "object"
            error:
              type: "string"
              example: "Something wemt wrong."
          examples:
            application/json: {"error": "Something went wrong."}
  /users/login:
    post:
      tags:
        - Users
      summary: "Starts login procedure for a user."
      description: "Log in a user. The parameters email and password can be specified via the form or via the query, but both need to be present."
      produces:
        - "application/json"
      parameters:
      - name: "email"
        description: "The email address of the user."
        in: "query"
        required: true
        example: "breyten@openstate.eu"
      - name: "password"
        description: "The password used for logging in."
        in: "query"
        required: true
        example: "password123"
      - name: "email"
        description: "The email address of the user."
        in: "form"
        required: true
        example: "breyten@openstate.eu"
      - name: "password"
        description: "The password used for logging in."
        in: "form"
        required: true
        example: "password123"
      responses:
        200:
          description: "The login was successful."
          schema:
            $ref: '#/definitions/User'
          examples:
            application/json:
              applicationId: "6e11ceb9-2b26-4c40-b399-182d4f9b97b6"
              email: "bje@dds.nl"
              email_verified: false
              family_name: "Ernsting"
              given_name: "Breyten"
              roles:
                - User
              sub: "6e11ceb9-2b26-4c40-b399-182d4f9b97b6"
        400:
          description: "Something went wrong logging in"
          schema:
            type: "object"
            error:
              type: "string"
          examples:
            application/json: {"error": "Something went wrong."}
    get:
      tags:
        - Users
      summary: "Starts login procedure for a user."
      description: "Log in a user. The parameters email and password can be specified via the form or via the query, but both need to be present."
      produces:
        - "application/json"
      parameters:
      - name: "email"
        description: "The email address of the user."
        in: "query"
        required: true
        example: "breyten@openstate.eu"
      - name: "password"
        description: "The password used for logging in."
        in: "query"
        required: true
        example: "password123"
      - name: "email"
        description: "The email address of the user."
        in: "form"
        required: true
        example: "breyten@openstate.eu"
      - name: "password"
        description: "The password used for logging in."
        in: "form"
        required: true
        example: "password123"
      responses:
        200:
          description: "The login was successful."
          schema:
            $ref: '#/definitions/User'
          examples:
            application/json:
              applicationId: "6e11ceb9-2b26-4c40-b399-182d4f9b97b6"
              email: "bje@dds.nl"
              email_verified: false
              family_name: "Ernsting"
              given_name: "Breyten"
              roles:
                - User
              sub: "6e11ceb9-2b26-4c40-b399-182d4f9b97b6"
        400:
          description: "Something went wrong logging in"
          schema:
            type: "object"
            error:
              type: "string"
          examples:
            application/json: {"error": "Something went wrong."}
  /users/passwordless/start:
    get:
      tags:
        - Users
      summary: "Starts a passwordless login for a user."
      description: "Starts a passwordless login for a user. An email address should be provided, or a user id."
      produces:
           - "application/json"
      parameters:
      - name: "email"
        description: "The email address of the user."
        in: "query"
        required: false
        example: "breyten@openstate.eu"
      - name: "user_id"
        description: "The user id."
        in: "query"
        required: false
        example: 1755
      responses:
        200:
          description: "The procedure was started"
          schema:
            type: "object"
            code:
              type: "string"
              description: "One time use login code. If this key is pesent, the passwordless login start was sucessful."
            error:
              type: "string"
              description: "If this key is present, there as an error sending the password login start. It describes the error."
          examples:
            application/json: {"code": "123456", "error": "Something went wrong."}
  /users/passwordless/complete:
    get:
      tags:
        - Users
      summary: "Completes a pasdswordless login for a user"
      description: "Complete the passwordless login. Places cookie and adds session info for further use."
      produces:
        - "application/json"
      parameters:
      - name: "code"
        description: "Login code that was sent."
        in: "query"
        required: true
        example: 123456
      responses:
        301:
          description: "The login was succesful. Redirect to front page."
        200:
          description: "Something went wrong logging in"
          schema:
            type: "object"
            error:
              description: "The contents describe the error encountered."
              type: "string"
          examples:
            application/json: {"error": "Something went wrong."}
  /users/register:
    post:
      tags:
        - Users
      summary: "Registers a new user"
      description: "Registeres a new user based on the given email address and password. A welcome email is sent."
      produces:
           - "application/json"
      parameters:
      - name: "email"
        description: "The email address of the user."
        in: "form"
        required: true
        example: breyten@openstate.eu
      - name: "password"
        description: "The password used for logging in."
        in: "form"
        required: true
        example: password123
      responses:
        200:
          description: "The registration was successful and an email was sent."
          schema:
            $ref: "#/definitions/User"
          examples:
            application/json:
              applicationId: "6e11ceb9-2b26-4c40-b399-182d4f9b97b6"
              email: "bje@dds.nl"
              email_verified: false
              family_name: "Ernsting"
              given_name: "Breyten"
              roles:
                - User
              sub: "6e11ceb9-2b26-4c40-b399-182d4f9b97b6"
        400:
          description: "Something went wrong registering the user. The contents describe the error encountered."
          schema:
            type: "object"
            error:
              type: "string"
          examples:
            application/json: {"error": "Something went wrong."}
  /users/verify:
    get:
      tags:
        - Users
      summary: "Verifies a user"
      description: "Verifies the e-mail address of a user. This is used to verify the user after they have registered."
      produces:
        - "application/json"
      parameters:
      - name: "id"
        description: "The token provided in the registration e-mail."
        in: "query"
        required: true
        example: "6e11ceb9-2b26-4c40-b399-182d4f9b97b6"
      responses:
        301:
          description: "The user was verified successfully. Redirect to front page."
        400:
          description: "Something went wrong verifying the user."
          schema:
            type: "object"
            error:
              type: "string"
          examples:
            application/json: {"error": "Something went wrong."}
  /users/logout:
    post:
      tags:
        - Users
      summary: "Logs a user out."
      description: "Removes a users session to log them out."
      produces:
        - "application/json"
      responses:
        200:
          description: "The user was logged out successfully."
          schema:
            type: "object"
            success:
              type: "string"
          examples:
            application/json: {"success":"true"}
        400:
          description: "Something went wrong logging the user out."
          schema:
            type: "object"
            error:
              type: "string"
              example: "Something wemt wrong."
          examples:
            application/json: {"error": "Something went wrong."}
...
  /archive/warc/create:
    get:
      tags:
        - WARC
      summary: "Starts a procedure to get heritrix to make a WARC archive of the given URL."
      description: "Instructs the heritrix backend to start a job to create a warc file of the given url. This happens asynchronously. A valid login is required for this to work."
      produces:
           - "application/json"
      parameters:
      - name: "url"
        description: "The website which is the subject of the warc file."
        in: "query"
        required: true
        example: https://openstate.eu/
      responses:
        200:
          description: "The creating of a warc file was started."
          schema:
            type: "object"
            job_id:
              type: "string"
              description: The archive id used if further requests
          examples:
            application/json:
              job_id: "123456-xx-123456-yy"
  /archive/warc/download/{archive_id}:
    get:
      tags:
        - WARC
      summary: "Downloads the given WARC file to your computer."
      produces:
           - "application/json"
      parameters:
      - name: "archive_id"
        in: path
        required: true
        example: "123456-xx-123456-yy"
      responses:
        301:
          description: "Starts downloading the file."
...
  /documents/download/{source}/{<external_item_id}:
    get:
      tags:
        - Documents
      summary: "Starts downloading an external file"
      parameters:
      - name: source
        description: specifies the source of the item (poliflw, obv, cvdr, openspending)
        in: path
        required: true
        example: cvdr
      - name: external_id
        description: "The external identifier for the file to be downloaded"
        in: path
        required: true
        example: 123456
      responses:
        301:
          description: "Redirects to the location of the file that is consequentially downloaded."
...
  /subscriptions/delete:
    get:
      tags:
        - Subscriptions
      summary: "Deletes an email subscription"
      description: "Deletes an email subscription. It uses the <a href=\"https://github.com/openstate/binoas/wiki/API\">binoas API</a> to do so."
      produces:
           - "application/json"
      parameters:
      - name: user_id
        in: query
        required: true
        description: "The id of the user."
        example: 1755
      - name: query_id
        in: query
        required: true
        description: "The id of the registered query, as previously gotten via a new subscription"
      responses:
        200:
          description: "The subscription was deleted."
          examples:
            application/json:
              status: "ok"
...
definitions:
...
  User:
    type: "object"
    applicationId:
      type: "string"
      description: "The application id."
    email:
      type: "string"
      description: "The email address of the user."
    email_verified:
      type: "boolean"
      description: "Whether the user was verified or not."
    family_name:
      type: "string"
      description: "The family name of the user."
    given_name:
      type: "string"
      description: "The given name of the user."
    roles:
      type: "list"
      description: "The roles that this user has in admin."
    sub:
      type: "string"
      description : "The interal id user for the user."